requirements.txt
azure-functions
github-copilot-sdk
